immob_backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-redis==5.6.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.14
redis==5.2.0
inflection==0.5.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.10
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.11.0
gunicorn==21.2.0

# Security packages
django-ratelimit==4.1.0
django-csp==1.2
bleach==6.2.0

